requirements.txt
flask
fastapi

sqlmodel

bcrypt
bleach
python-multipart
python-slugify
fastapi_login
flask_babel

sentry_sdk
pytest

requests
beautifulsoup4

uvicorn
